#4 Layout and blocks: fixed: moving controllers and page layout and templares
