+++ app/code/Oleksandrb/ControllerDemos/Observer/DumpMergedLayoutAndHandles.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Oleksandrb\LayoutDebug\Observer;
+namespace Oleksandrb\ControllerDemos\Observer;
 
 use Magento\Framework\App\Filesystem\DirectoryList;
 use Magento\Framework\Event\Observer;
