+++ app/code/Oleksandrb/ControllerDemos/etc/frontend/routes.xml
@@ -10,5 +10,5 @@
 </config>
 
 <!--
-https://oleksandr-babii-magento.local/oleksandrb-controllers-demo/foobar_yetanotherfolder/rawresponsedemo
+https://oleksandr-babii-magento.local/oleksandrb-controllers-demo/foobar_yetanotherfolder/responseform
 -->
