+++ app/code/Oleksandrb/ControllerDemos/view/frontend/layout/oleksandrb_controllersdemo_foobar_yetanotherfolder_pageform.xml
@@ -5,10 +5,10 @@
 >
     <body>
         <referenceContainer name="content">
-            <container name="oleksandrb.layout.debug.container" htmlTag="div" htmlClass="oleksandrb-layout-debug">
+            <container name="oleksandrb.controller.demos.container" htmlTag="div" htmlClass="oleksandrb-controller-demos">
                 <block class="Magento\Framework\View\Element\Template"
-                       name="oleksandrb.layout.debug.text.block.with.template"
-                       template="Oleksandrb_LayoutDebug::demo.phtml">
+                       name="oleksandrb.controller.demos.text.block.with.template"
+                       template="Oleksandrb_ControllerDemos::form.phtml">
                 </block>
             </container>
         </referenceContainer>
