--- app/code/Oleksandrb/LayoutDebug/etc/frontend/routes.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:noNamespaceSchemaLocation="urn:magento:framework:App/etc/routes.xsd"
->
-    <router id="standard">
-        <route id="oleksandrb_layout_debug" frontName="oleksandrb-layout-debug">
-            <module name="Oleksandrb_LayoutDebug"/>
-        </route>
-    </router>
-</config>
-
