--- app/code/Oleksandrb/LayoutDebug/registration.php
@@ -1,9 +1,0 @@
-<?php
-
-use Magento\Framework\Component\ComponentRegistrar;
-
-ComponentRegistrar::register(
-    ComponentRegistrar::MODULE,
-    'Oleksandrb_LayoutDebug',
-    __DIR__
-);
